Step 1, separate OS implementations
